Synopsys patch to support EV processors

+++ 2.4.9/modules/calib3d/test/test_reproject_image_to_3d.cpp
@@ -137,7 +137,7 @@ protected:
             {
                 InT d = disp(y, x);
 
-                double from[4] = { x, y, d, 1 };
+                double from[4] = { (double)x, (double)y, (double)d, (double)1 };
                 Mat_<double> res = Q * Mat_<double>(4, 1, from);
                 res /= res(3, 0);
